Add higher resolutions to windows ico files
Ad the resolution of 512x512 to the windows icon. On desktops with higher zoom value, this helps to see the app icon in full size - instead of a small icon in a gray box. To be prepared for the future, the shield.ico in the asset also includes 768 x 768. The systray icon is now smaller and just includes the resolutions 16, 32 and 48.